[ag] #131 - simplify tests - centralize test types
