+++ tests/ag/test.cpp
@@ -8,7 +8,7 @@
 #endif
 
 #include "includes/tests/ag/details/mp.hpp"
-#include "includes/tests/ag/conversion/to_tuple.hpp"    // TODO: to_tuple_t vs. expected !!!
+#include "includes/tests/ag/conversion/to_tuple.hpp"
 #include "includes/tests/ag/conversion/to_tuple_view.hpp"
 #include "includes/tests/ag/size.hpp"
 // #include "includes/tests/ag/element.hpp" // WIP
